requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
PycordLib==1.2.0
requests==2.18.1
schedule==0.4.3
six==1.10.0
urllib3==1.21.1
websocket-client==0.44.0
